python3/requirements.txt
gql[aiohttp]
python-dotenv
